+++ AesNi/Aes256Key.cs
@@ -37,63 +37,63 @@ namespace AesNi
             ref var keyRef = ref MemoryMarshal.GetReference(key);
 
             var tmp1 = ReadUnaligned(ref keyRef);
-            var tmp3 = ReadUnalignedOffset(ref keyRef, (IntPtr) (1 * BytesPerRoundKey));
+            var tmp3 = ReadUnalignedOffset(ref keyRef, 1 * BytesPerRoundKey);
 
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (0 * BytesPerRoundKey), tmp1);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (1 * BytesPerRoundKey), tmp3);
+            WriteUnalignedOffset(ref expandedKey, 0 * BytesPerRoundKey, tmp1);
+            WriteUnalignedOffset(ref expandedKey, 1 * BytesPerRoundKey, tmp3);
 
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (27 * BytesPerRoundKey), InverseMixColumns(tmp3));
+            WriteUnalignedOffset(ref expandedKey, 27 * BytesPerRoundKey, InverseMixColumns(tmp3));
 
             Aes256KeyExp1(ref tmp1, tmp3, 0x01);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (2 * BytesPerRoundKey), tmp1);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (26 * BytesPerRoundKey), InverseMixColumns(tmp1));
+            WriteUnalignedOffset(ref expandedKey, 2 * BytesPerRoundKey, tmp1);
+            WriteUnalignedOffset(ref expandedKey, 26 * BytesPerRoundKey, InverseMixColumns(tmp1));
 
             Aes256KeyExp2(tmp1, ref tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (3 * BytesPerRoundKey), tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (25 * BytesPerRoundKey), InverseMixColumns(tmp3));
+            WriteUnalignedOffset(ref expandedKey, 3 * BytesPerRoundKey, tmp3);
+            WriteUnalignedOffset(ref expandedKey, 25 * BytesPerRoundKey, InverseMixColumns(tmp3));
 
             Aes256KeyExp1(ref tmp1, tmp3, 0x02);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (4 * BytesPerRoundKey), tmp1);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (24 * BytesPerRoundKey), InverseMixColumns(tmp1));
+            WriteUnalignedOffset(ref expandedKey, 4 * BytesPerRoundKey, tmp1);
+            WriteUnalignedOffset(ref expandedKey, 24 * BytesPerRoundKey, InverseMixColumns(tmp1));
 
             Aes256KeyExp2(tmp1, ref tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (5 * BytesPerRoundKey), tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (23 * BytesPerRoundKey), InverseMixColumns(tmp3));
+            WriteUnalignedOffset(ref expandedKey, 5 * BytesPerRoundKey, tmp3);
+            WriteUnalignedOffset(ref expandedKey, 23 * BytesPerRoundKey, InverseMixColumns(tmp3));
 
             Aes256KeyExp1(ref tmp1, tmp3, 0x04);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (6 * BytesPerRoundKey), tmp1);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (22 * BytesPerRoundKey), InverseMixColumns(tmp1));
+            WriteUnalignedOffset(ref expandedKey, 6 * BytesPerRoundKey, tmp1);
+            WriteUnalignedOffset(ref expandedKey, 22 * BytesPerRoundKey, InverseMixColumns(tmp1));
 
             Aes256KeyExp2(tmp1, ref tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (7 * BytesPerRoundKey), tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (21 * BytesPerRoundKey), InverseMixColumns(tmp3));
+            WriteUnalignedOffset(ref expandedKey, 7 * BytesPerRoundKey, tmp3);
+            WriteUnalignedOffset(ref expandedKey, 21 * BytesPerRoundKey, InverseMixColumns(tmp3));
 
             Aes256KeyExp1(ref tmp1, tmp3, 0x08);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (8 * BytesPerRoundKey), tmp1);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (20 * BytesPerRoundKey), InverseMixColumns(tmp1));
+            WriteUnalignedOffset(ref expandedKey, 8 * BytesPerRoundKey, tmp1);
+            WriteUnalignedOffset(ref expandedKey, 20 * BytesPerRoundKey, InverseMixColumns(tmp1));
 
             Aes256KeyExp2(tmp1, ref tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (9 * BytesPerRoundKey), tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (19 * BytesPerRoundKey), InverseMixColumns(tmp3));
+            WriteUnalignedOffset(ref expandedKey, 9 * BytesPerRoundKey, tmp3);
+            WriteUnalignedOffset(ref expandedKey, 19 * BytesPerRoundKey, InverseMixColumns(tmp3));
 
             Aes256KeyExp1(ref tmp1, tmp3, 0x10);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (10 * BytesPerRoundKey), tmp1);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (18 * BytesPerRoundKey), InverseMixColumns(tmp1));
+            WriteUnalignedOffset(ref expandedKey, 10 * BytesPerRoundKey, tmp1);
+            WriteUnalignedOffset(ref expandedKey, 18 * BytesPerRoundKey, InverseMixColumns(tmp1));
 
             Aes256KeyExp2(tmp1, ref tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (11 * BytesPerRoundKey), tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (17 * BytesPerRoundKey), InverseMixColumns(tmp3));
+            WriteUnalignedOffset(ref expandedKey, 11 * BytesPerRoundKey, tmp3);
+            WriteUnalignedOffset(ref expandedKey, 17 * BytesPerRoundKey, InverseMixColumns(tmp3));
 
             Aes256KeyExp1(ref tmp1, tmp3, 0x20);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (12 * BytesPerRoundKey), tmp1);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (16 * BytesPerRoundKey), InverseMixColumns(tmp1));
+            WriteUnalignedOffset(ref expandedKey, 12 * BytesPerRoundKey, tmp1);
+            WriteUnalignedOffset(ref expandedKey, 16 * BytesPerRoundKey, InverseMixColumns(tmp1));
 
             Aes256KeyExp2(tmp1, ref tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (13 * BytesPerRoundKey), tmp3);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (15 * BytesPerRoundKey), InverseMixColumns(tmp3));
+            WriteUnalignedOffset(ref expandedKey, 13 * BytesPerRoundKey, tmp3);
+            WriteUnalignedOffset(ref expandedKey, 15 * BytesPerRoundKey, InverseMixColumns(tmp3));
 
             Aes256KeyExp1(ref tmp1, tmp3, 0x40);
-            WriteUnalignedOffset(ref expandedKey, (IntPtr) (14 * BytesPerRoundKey), tmp1);
+            WriteUnalignedOffset(ref expandedKey, 14 * BytesPerRoundKey, tmp1);
         }
 
         // https://www.intel.com/content/dam/doc/white-paper/advanced-encryption-standard-new-instructions-set-paper.pdf
